pyproject.toml
[project]
name = "snake"
version = "0.1.0"
description = "snake"
authors = [
    {name = "Aearsears",email = "53102185+Aearsears@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce (>=2.5.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
